+++ kern/include/syscall.h
@@ -59,7 +59,6 @@ void enter_forked_process(struct trapframe *tf);
 __DEAD void enter_new_process(int argc, userptr_t argv, userptr_t env,
 		       vaddr_t stackptr, vaddr_t entrypoint);
 
-
 /*
  * Prototypes for IN-KERNEL entry points for system call implementations.
  */
@@ -71,7 +70,7 @@ int sys_read(int fd, void *buf,size_t buflen,int *retval);
 int sys_write(int fd, const void *buf, size_t buflen,int *retval);
 int sys_close(int fd);
 // int sys_lseek(int fd, off_t pos, int whence,int *retval,int *retval_1);
-int sys_lseek(int fd, off_t pos, int whence,int *retval,int *retval_1, uint32_t new_position);
+int sys_lseek(int fd, off_t pos, int whence,int *retval,int *retval_1, uint64_t new_position);
 int sys_dup2(int old_fd, int new_fd,int *retval);
 int sys_chdir(userptr_t pathname,int *retval);
 int sys___getcwd(char *fname,size_t buflen, int *retval);
